IVQ-CBM/requirements.txt
einops
matplotlib
open-clip-torch
PyYAML
scikit-image
scikit-learn
scipy
tensorboard
timm
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
nvitop